+++ src/main/java/com/project/pillpal/security/jwt/JwtService.java
@@ -1,6 +1,6 @@
-package com.project.pillpal.user.security.jwt;
+package com.project.pillpal.security.jwt;
 
-import com.project.pillpal.user.security.UserDetail;
+import com.project.pillpal.security.UserDetail;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.io.Decoders;
