+++ src/main/java/com/project/pillpal/user/Auth/AuthController.java
@@ -3,8 +3,8 @@ package com.project.pillpal.user.Auth;
 import com.project.pillpal.user.dtos.JwtResponse;
 import com.project.pillpal.user.dtos.UserRequest;
 import com.project.pillpal.user.dtos.UserResponse;
-import com.project.pillpal.user.security.UserDetail;
-import com.project.pillpal.user.security.jwt.JwtService;
+import com.project.pillpal.security.UserDetail;
+import com.project.pillpal.security.jwt.JwtService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
